refactor(wsl2d): isolate io_uring behind wrapper crate (#3)
