+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "3"
-members = ["crates/ramshared-cli", "crates/ramshared-tier", "crates/ramshared-cuda", "crates/ramshared-block", "crates/ramshared-wsl2d", "crates/ramshared-integrity"]
+members = ["crates/ramshared-cli", "crates/ramshared-tier", "crates/ramshared-cuda", "crates/ramshared-block", "crates/ramshared-wsl2d", "crates/ramshared-integrity", "crates/ramshared-uring"]
 
 [workspace.package]
 edition = "2024"
