+++ crates/ramshared-uring/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ramshared-uring"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[lib]
+name = "ramshared_uring"
+path = "src/lib.rs"
+
+[dependencies]
+io-uring = "0.7.12"
+
+[lints.clippy]
+unwrap_used = "deny"
+expect_used = "deny"
